+++ lib/trace_segmentation.py
@@ -274,7 +274,7 @@ def get_ridge_line(ridges_h, ridges_v, region):
   return ridge_line
 
 def get_image_values(img_gray, coords):
-  return map(lambda pt : int(255*img_gray[tuple(pt)]), coords)
+  return map(lambda pt : int(255*img_gray[tuple(pt.astype(int))]), coords)
 
 def get_ridge_coords_in_region(ridges, coord_list):
   '''
